COZMO-1622 Docking Test Behavior (#1037)
* COZMO-1622 Docking Test Behavior

Simple behavior for testing docking. Saves information for each dock attempt to log as well as images from robot.

+++ basestation/src/anki/cozmo/basestation/actions/dockActions.cpp
@@ -26,7 +26,7 @@ namespace Anki {
   namespace Cozmo {
   
     // Which docking method actions should use
-    CONSOLE_VAR(u8, kDockingMethod, "Docking", (u8)DockingMethod::BLIND_DOCKING);
+    CONSOLE_VAR(u32, kDockingMethod, "Docking", (u8)DockingMethod::BLIND_DOCKING);
     
     // Helper function for computing the distance-to-preActionPose threshold,
     // given how far robot is from actionObject
